chore(package): refactor remove unnecessary "index.js" module.
The "main" keyword in package.json takes care of pointing to the entry point to the module.

It was previously pointing to a non-existent file and thus not working.

"main" was updated to point to the true entry point in ./lib/index.js,
rendering the shell file of "./index.js" unnecessary.